requirements.txt
agate==1.9.1
annotated-types==0.7.0
appdirs==1.4.4
asttokens==2.4.1
attrs==23.2.0
babel==2.16.0
beautifulsoup4==4.12.3
blinker==1.8.2
Bottleneck==1.4.0
cachetools==5.4.0
cattrs==23.2.3
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
daff==1.3.46
db-dtypes==1.2.0
dbt-adapters==1.6.1
dbt-bigquery==1.8.2
dbt-common==1.8.0
dbt-core==1.8.6
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.5.1
debugpy==1.8.2
decorator==5.1.1
deepdiff==7.0.1
Deprecated==1.2.14
diff_cover==9.2.0
exceptiongroup==1.2.2
executing==2.0.1
Flask==3.0.3
fonttools==4.53.1
frozendict==2.4.4
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-appengine-logging==1.4.5
google-cloud-audit-log==0.2.5
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.25.0
google-cloud-core==2.4.1
google-cloud-dataproc==5.12.0
google-cloud-logging==3.11.1
google-cloud-secret-manager==2.20.2
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.65.4
grpcio-status==1.62.2
html5lib==1.1
httplib2==0.22.0
idna==3.7
importlib-metadata==6.11.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
leather==0.4.0
llvmlite==0.43.0
Logbook==1.5.3
lxml==5.2.2
MarkupSafe==2.1.5
mashumaro==3.13.1
matplotlib==3.9.1
matplotlib-inline==0.1.7
minimal-snowplow-tracker==0.0.2
more-itertools==10.5.0
msgpack==1.1.0
multitasking==0.0.11
nest-asyncio==1.6.0
networkx==3.3
numba==0.60.0
numexpr==2.10.1
numpy==2.0.1
opentelemetry-api==1.26.0
ordered-set==4.1.0
packaging==24.1
pandas==2.2.2
parsedatetime==2.6
parso==0.8.4
pathspec==0.12.1
peewee==3.17.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.2
pyrate-limiter==2.10.0
pytest==8.3.2
pytest-flask==1.3.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.0.3
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-cache==1.2.1
requests-mock==1.12.1
requests-ratelimiter==0.7.0
rpds-py==0.20.0
rsa==4.9
ruff==0.5.7
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
soupsieve==2.5
sqlfluff==3.2.0
sqlparse==0.5.1
stack-data==0.6.3
tblib==3.0.0
text-unidecode==1.3
threadpoolctl==3.5.0
toml==0.10.2
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
wrapt==1.16.0
xarray==2024.6.0
yfinance==0.2.41
zipp==3.19.2
